Add files via upload
This Excel file contains the modified CSV file as well as a pivot table and graph displaying long-term growth in cryptocurrencies. 

This will give key insight into the process I took to answer the first of three questions I posed in the final research report that will be at the end of this repository.