Add files via upload
This python file addresses the second question poised in the project report, "Can the Gains and Losses of Bitcoin be predicted by a machine algorithm?". Throughout the notebook scikit's k nearest neighbors algorithm will be used to attempt to predict the gains or losses of bitcoin on any given day. Afterwards hyperparamater training is used to improve the results.