Add files via upload
This Python notebook goes over the 3rd question asked in the report, "Which cryptocurrencies are the most volatile?". By using a box plot to test the spread in the daily change in price of three popular crypto coins. This will give a gauge on the volatility on these three coins.

Also included in this notebook are a couple of other graphs and questions that are analyzed, that were going to be potentially used for the project.